backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.18
sqlalchemy==2.0.36
